backend/config/requirements.txt
fastapi
uvicorn[standard]
ragflow-sdk
python-dotenv